Completely parallelize splitting operations
[ ENHANCEMENT ]  Each file that is split is now done in parallel.
This differs from before, where each directory (that had files)
were done in parallel

[ FIXED ]  Rewrote how to process each file to split from a given
CUE file.  Now we can handle filenames that are specified with one
or more extensions that may NOT be in the current directory, but
are intended to be split (based off the file's basename)

[ FIXED ]  Shntools does NOT add a period before the extension IF
the file ends with a period.  Now, we check for a period and add
the period ourselves if need be

[ FIXED ]  Total files found in current directory with multiple files

[ FIXED ]  Correctly display current file that's being split

[ FIXED ]  Correctly handle the UI count displayed